refactor(nms): Separate the ApnContext
