+++ nms/app/views/traffic/__tests__/TrafficOverviewTest.tsx
@@ -10,7 +10,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import ApnContext from '../../../components/context/ApnContext';
 import MagmaAPI from '../../../api/MagmaAPI';
 import MuiStylesThemeProvider from '@material-ui/styles/ThemeProvider';
 import PolicyContext, {
@@ -19,23 +18,19 @@ import PolicyContext, {
 import React from 'react';
 import TrafficDashboard from '../TrafficOverview';
 import defaultTheme from '../../../theme/default';
+import {ApnContextProvider} from '../../../components/context/ApnContext';
 
-import {
-  Apn,
-  PolicyQosProfile,
-  PolicyRule,
-  RatingGroup,
-} from '../../../../generated';
 import {MemoryRouter, Route, Routes} from 'react-router-dom';
 import {MuiThemeProvider} from '@material-ui/core/styles';
 import {PolicyId} from '../../../../shared/types/network';
-import {SetApnState} from '../../../state/lte/ApnState';
+import {PolicyQosProfile, PolicyRule, RatingGroup} from '../../../../generated';
 import {
   SetPolicyState,
   SetQosProfileState,
   SetRatingGroupState,
 } from '../../../state/PolicyState';
 import {fireEvent, render, wait} from '@testing-library/react';
+import {mockAPI} from '../../../util/TestUtils';
 
 jest.mock('axios');
 jest.mock('../../../../app/hooks/useSnackbar');
@@ -134,6 +129,10 @@ const ratingGroups: Record<string, RatingGroup> = {
 };
 
 describe('<TrafficDashboard />', () => {
+  beforeEach(() => {
+    mockAPI(MagmaAPI.apns, 'lteNetworkIdApnsGet', apns);
+  });
+
   const networkId = 'test';
   const policyCtx: PolicyContextType = {
     state: policies,
@@ -171,19 +170,6 @@ describe('<TrafficDashboard />', () => {
     refetch: () => {},
   };
 
-  const apnCtx = {
-    state: apns,
-    setState: (key: string, value?: Apn) => {
-      return SetApnState({
-        apns,
-        setApns: () => {},
-        networkId,
-        key,
-        value,
-      });
-    },
-  };
-
   const Wrapper = () => (
     <MemoryRouter
       initialEntries={['/nms/test/traffic/policy']}
@@ -191,14 +177,14 @@ describe('<TrafficDashboard />', () => {
       <MuiThemeProvider theme={defaultTheme}>
         <MuiStylesThemeProvider theme={defaultTheme}>
           <PolicyContext.Provider value={policyCtx}>
-            <ApnContext.Provider value={apnCtx}>
+            <ApnContextProvider networkId={networkId}>
               <Routes>
                 <Route
                   path="/nms/:networkId/traffic/*"
                   element={<TrafficDashboard />}
                 />
               </Routes>
-            </ApnContext.Provider>
+            </ApnContextProvider>
           </PolicyContext.Provider>
         </MuiStylesThemeProvider>
       </MuiThemeProvider>
@@ -373,13 +359,15 @@ describe('<TrafficDashboard />', () => {
 describe('<TrafficDashboard APNs/>', () => {
   const {location} = window;
 
-  beforeAll((): void => {
+  beforeEach((): void => {
     window.location = {
       pathname: '/nms/test/traffic/apn',
     } as Location;
+
+    mockAPI(MagmaAPI.apns, 'lteNetworkIdApnsGet', apns);
   });
 
-  afterAll((): void => {
+  afterEach((): void => {
     window.location = location;
   });
 
@@ -403,31 +391,19 @@ describe('<TrafficDashboard APNs/>', () => {
     },
     refetch: () => {},
   };
-  const apnCtx = {
-    state: apns,
-    setState: (key: string, value?: Apn) => {
-      return SetApnState({
-        apns,
-        setApns: () => {},
-        networkId,
-        key,
-        value,
-      });
-    },
-  };
   const Wrapper = () => (
     <MemoryRouter initialEntries={['/nms/test/traffic/apn']} initialIndex={0}>
       <MuiThemeProvider theme={defaultTheme}>
         <MuiStylesThemeProvider theme={defaultTheme}>
           <PolicyContext.Provider value={policyCtx}>
-            <ApnContext.Provider value={apnCtx}>
+            <ApnContextProvider networkId={networkId}>
               <Routes>
                 <Route
                   path="/nms/:networkId/traffic/*"
                   element={<TrafficDashboard />}
                 />
               </Routes>
-            </ApnContext.Provider>
+            </ApnContextProvider>
           </PolicyContext.Provider>
         </MuiStylesThemeProvider>
       </MuiThemeProvider>
